+++ src/libs/blueprint/conduit_blueprint_mesh_utils.hpp
@@ -453,14 +453,20 @@ namespace topology
     //-------------------------------------------------------------------------
     namespace unstructured
     {
-        // TODO(JRC): Expose this 'cache' version of the function publicly?
         //-------------------------------------------------------------------------
-        void CONDUIT_BLUEPRINT_API generate_offsets(Node &topo,
+        // Generates element offsets for given topo
+        void CONDUIT_BLUEPRINT_API generate_offsets(const Node &topo,
                                                     Node &dest);
 
         //-------------------------------------------------------------------------
+        // Generates element and subelement offsets for given topo
         void CONDUIT_BLUEPRINT_API generate_offsets(const Node &topo,
-                                                    Node &dest);
+                                                    Node &dest_ele_offsets,
+                                                    Node &dest_subele_offsets);
+
+        //-------------------------------------------------------------------------
+        // Adds offsets to given topo
+        void CONDUIT_BLUEPRINT_API generate_offsets_inline(Node &topo);
 
         //-------------------------------------------------------------------------
         std::vector<index_t> CONDUIT_BLUEPRINT_API points(const Node &topo,